+++ src/components/common/BottomTabs.tsx
@@ -8,7 +8,7 @@ import {COLOR} from '@src/constants/color';
 import {HomeScreen} from '@src/screens/HomeScreen';
 import {TourScreen} from '@src/screens/tour/TourScreen';
 import {DiaryScreen} from '@src/screens/diary/DiaryScreen';
-import {MypageScreen} from '@src/screens/MypageScreen';
+import {MypageScreen} from '@src/screens/mypage/MypageScreen';
 import {DiscoveryScreen} from '@src/screens/DiscoveryScreen';
 
 interface IconProps {
